% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SanBubble.R
\name{SanBubble}
\alias{SanBubble}
\title{桑基气泡图绘图工具}
\usage{
SanBubble(Bubble, Sankey, watermark)
}
\arguments{
\item{Bubble}{气泡图数据框}

\item{Sankey}{桑基图数据库}

\item{watermark}{YES是有水印,NO是无水印}
}
\description{
桑基气泡图绘图工具
}
